adding the new architecture for quality measures
